requirements.txt
flask>=2.3,<3.1
opencv-python
numpy>=1.24.3
ultralytics>=8.0.0
pillow>=10.0.0
